+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 0.3.0
+## [Unreleased] - 0.3.1
+
+### Added
+-
+
+### Fixed
+-
+
+### Changed
+-
+
+### Removed
+-
+
+## [0.3.0] - 2023-10-14
 
 ### Added
 - `Regex` scanner to the prompt
@@ -13,9 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `JSON` output scanner
 - Best practices to the documentation
 - `LanguageSame` output scanner to check that the prompt and output languages are the same
-
-### Fixed
--
 
 ### Changed
 - `BanSubstrings` can match all substrings in addition to any of them
@@ -188,6 +199,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [Toxicity](./llm_guard/output_scanners/toxicity.py)
 
 [Unreleased]: https://github.com/laiyer-ai/llm-guard/commits/main
+[0.3.0]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.3.0
 [0.2.4]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.2.4
 [0.2.3]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.2.3
 [0.2.2]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.2.2